pyproject.toml
[project]
name = "fractal-hunter-dataset"
version = "0.1.0"
description = "Synthetic dataset for identity recognition under structured distortion"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.0",
    "pandas>=3.0.2",
    "pillow>=11.0.0",
]
